+++ packages/react-native-contentpass/src/contentpassTokenUtils/fetchContentpassToken.test.ts
@@ -29,6 +29,7 @@ describe('fetchContentpassToken', () => {
         headers: {
           'Content-Type': 'application/x-www-form-urlencoded',
         },
+        signal: expect.any(AbortSignal),
       }
     );
   });